Frontend: Known bug fixed
-Bug Fix Description:

  -Bug Overview:

The bug occurred when navigating from a user's profile (who is not the current user) to the current user's profile. The data didn't load properly during this transition.
-Bug Manifestation:

When navigating between a user profile and the current user's profile, the data wasn't displayed as expected.
-Temporary Fix:

The issue was resolved by reloading the entire component using window.location.reload on the NavigationEnd event.
-Limitation:

While this approach fixes the bug, it is not an optimal solution because it reloads the whole page, which can lead to performance issues and a less smooth user experience.